New ability Witch Hunt
Added new ability Witch Hunt, witch (ahah see, this is  a pun) behaves the opposite way than scorch, it destroys the lowest cards in the opposite row. Only a few specific cards have it. Let's be honnest, this is a muster/monster killer because it goes around the usual protections for muster cards