Bugfix, balance, ai-vs-ai and sim/game tracking addition
Significant update done for quite some time. Various bug fixes and attempts to rebalance some decks.
Added the possibility to track what going on during the game to perform analysis later.
Game-engine was reworked to be more flexible and now we can have AI vs AI games. From this, we could also get rid of the interface to simulate games in large numbers to for instance  study the balance between decks. This is accessible through the new sim.html page.